m4.xml:  The instance name is now different than the template name.
init file is added to resolve the import problems.

ta_helper:
The constraint is now a class variable (not global anymore).
NTA (the network of TA) is returned to and stored in Tamus. Thus, we do not read the input file for each "check". Instead, for each verifyta call, the TA is set (TA of interest), and the resulting file is stored in a file to feed to verifyta.

timed_automaton:
A bug in relaxed automaton generation is fixed. In particular, there was a problem when multiple constraints are removed from the same transition (or location) the relaxation result was incomplete (wrong), since only the first one was removed. The structure of the constraint registry and generate relaxed template  function are changed.

The transitions without guards (guard = "") was added to constraint registry. To avoid this, empty string check is added.

In this version, the input file is read only once, which reduces the computation time.