The analysis is now run on whole TA (instead of path analysis)
The single path optimization is removed. It can be used later, the related methods are still in timed_automata, ta_helper.

Note:
The analysis takes a long time on test_case2, this is due to the number of constraints. As all paths are simple we can improve computation time via path analysis.